@@ -1,3 +1,3 @@
 module github.com/mitranim/gax
 
-go 1.16
+go 1.20
